requirements.txt
Django==5.1.14
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
markdown==3.7
docker==7.1.0
gunicorn==23.0.0
uvicorn[standard]==0.34.0
django-mcp-server==0.1.0
channels==4.2.0
daphne==4.1.2
croniter==1.4.1

# Phase 3: RAG dependencies
sentence-transformers==2.2.2  # Local embedding model
pypdf==3.17.0  # PDF text extraction
python-docx==1.1.0  # DOCX text extraction

